Differentiate our own type and type from CMS.
