+++ src/components/Header.astro
@@ -1,5 +1,5 @@
 ---
-import type { HeaderOption } from "@/types";
+import type { HeaderOption } from "@/types/types";
 
 type HeaderProp = {
     activeNavigationOption : HeaderOption
