+++ src/layouts/Layout.astro
@@ -1,7 +1,7 @@
 ---
 import Footer from '@/components/Footer.astro';
 import Header from '@/components/Header.astro';
-import type { HeaderOption } from '@/types';
+import type { HeaderOption } from '@/types/types';
 
 export interface Props {
 	title: string;
