+++ src/types/cms.ts
@@ -26,13 +26,3 @@ export interface ProjectCardContent {
 export interface AllProjectCardContent {
     allProjects : ProjectCardContent[]
 }
-
-export interface Date {
-    date : number,
-    month : number,
-    year : number,
-}
-
-const headerOptions = ["home", "portofolio"] as const;
-
-export type HeaderOption = typeof headerOptions[number];
